requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
requests
django-redis
redis
python-dotenv
gunicorn
dj-database-url
psycopg2-binary
whitenoise
